Move PathWithLines to cucumber.runtime.model, where its user is
