+++ core/src/test/java/cucumber/runtime/model/PathWithLinesTest.java
@@ -1,4 +1,4 @@
-package cucumber.runtime;
+package cucumber.runtime.model;
 
 import org.junit.Test;
